Update genotypeio.py
As per my issue in https://github.com/broadinstitute/tensorqtl/issues/144 i sometimes experience freeze of tensorqtl as there is no variants in a range to test on a particular chromosome. MAF and window dropped all the genes to test on chromosome and because it still loops through it it freezes up. 

Here I define the chromosones to test only based on what actually is there to test.